+++ NEWS.md
@@ -1,7 +1,11 @@
 # shinyDiveR 0.2.0.0
 
-## Minor improvements and fixes
+## Major improvements
 
 * Added a `NEWS.md` file to track changes to the package.
+
+* Reactive language option
+
+## Minor improvements and fixes
 
 * Add style via a css file and formated the package following golem workflow.